(cleanup) Reorganize components and containers based on use and add comments to catagorize what our components are importing
